Keep track of FPGA time rollovers with 64-bit time.
This allows both greater than 72 minute (2^32 * 1 us) timeouts and also
gracefully handles notifiers across the FPGA time counter rollover.

+++ hal/include/HAL/HAL.hpp
@@ -222,7 +222,7 @@ extern "C"
 
 	uint16_t getFPGAVersion(int32_t *status);
 	uint32_t getFPGARevision(int32_t *status);
-	uint32_t getFPGATime(int32_t *status);
+	uint64_t getFPGATime(int32_t *status);
 
 	bool getFPGAButton(int32_t *status);
