+++ wpilibc/simulation/src/Utility.cpp
@@ -159,7 +159,7 @@ bool wpi_assertNotEqual_impl(int valueA,
  *
  * @return The current time in microseconds according to the FPGA (since FPGA reset).
  */
-uint32_t GetFPGATime()
+uint64_t GetFPGATime()
 {
 	return wpilib::internal::simTime * 1e6;
 }
